Hacker Earth Hackaton
Problem statement
Flying has been the go-to mode of travel for years now; it is time-saving, affordable, and extremely convenient. According to the FAA, 2,781,971 passengers fly every day in the US, as in June 2019. Passengers reckon that flying is very safe, considering strict inspections are conducted and security measures are taken to avoid and/or mitigate any mishappenings. However, there remain a few chances of unfortunate incidents.

Imagine you have been hired by a leading airline. You are required to build Machine Learning models to anticipate and classify the severity of any airplane accident based on past incidents. With this, all airlines, even the entire aviation industry, can predict the severity of airplane accidents caused due to various factors and, correspondingly, have a plan of action to minimize the risk associated with them.

Data set
The dataset consists of certain parameters recorded during the incident⁠ such as cabin temperature, turbulence experienced, number of safety complaints prior to the accident, number of days since the last inspection was conducted before the incident, an estimation of the pilot’s control given the various factors at play, and the likes.
The benefits of practicing this problem by using Machine Learning techniques are as follows:
This challenge will encourage you to apply your Machine Learning skills to build models that can anticipate the severity of any airplane accident
This challenge will help you enhance your knowledge of classification actively. Classification is one of the basic building blocks of Machine Learning
We challenge you to build a model that predicts how severe an airplane accident could be.